This is an outline for the trait prediction script.  It contains primarily library functions (the script functions will go in PICRUST/scripts/).
For now the functions don't take into account rate matrices, probabilities, or distributions of counts.

They are set up to take into account reconstructions from programs that can't handle amiguous characters, and because of that drop & prune the parent of the node that is to be reconstructed (the 'pruned parent problem')

I also added some capability to handle arrays where some values are missing.
This might be useful in cases where only certain traits are known to be present, and others could be either present or absent.  For example, an unsequenced genome where some genes of interest have been picked up by PCR.  

The algorithm accounts for the pruned parent problem by using a weighted average of the last ancestral state (usually the grandparent) and the sibling node(s) to predict the parent.   Basically this is very similar to the Josh et al exponential weighting model at short ranges, but informed by ancestral states being propagated up from below. 

The weight function is abstracted so we can play with various options.