Add support files for ASR methods ACE and Count.
ace.py and count.py application controllers will use these included programs. 
Note that Count software is included for ease of use during testing. 
If Count is chosen as a "good" ASR method then it may be removed and users would need to obtain it themselves and set a 'COUNT_JAR' environment variable.