Script for make_test_datasets.py currently this supports generation of a directory with many different trees.  Each tree is modified based on distance parameter.
For now this is mostly in support of validation for the genome predictions.   The metagenomic validation will use some of the same underlying functions.

The current interface is likely to change to support metagenomic as well as genomic validation

Available methods include:

-- Exclude tips by distance:

- remove all tips within d of our tip of interest.

- randomize labels for all tips within d of our tip of interest


For each method the script iterates over a range of distances.