Change options for calculating NSTI in predict_traits.py
If calculating NSTI, most of the time we don't need a separate file
with the measures since it is in the main output file (both tab or biom format),
so change -a option to a store_true option (instead of specifying a filepath).

Might still be convenient to get NSTI values without having to do all of the predict trait steps, so add option --output_accuracy_metric_results that takes a filepath and outputs NSTI to that file and exits before doing rest of predictions.