Run tests with pytest instead of all_tests
The all_tests.py was a semi-custom test runner, adapted from QIIME and
PyCogent, that ran the unit tests and some script tests. The script
tests depend on `qiime.test.run_script_usage_tests` from QIIME1 to run.
However, with support for QIIME1 now dropped (in favor of QIIME2), and
the increasing difficulty of installing QIIME1 alongside PICRUSt, it
seems like a reasonable time to move to a more generic test runner.

The test data in `picrust_test_data/` will not be used for now. A new
end-to-end test suite could be written in the future, perhaps leveraging
what's in use in QIIME2. Otherwise, explicit end-to-end tests aren't the
worst idea.